Fix leaked open connections by closing transport
Found SFTPClient.close() was not closing the underlying transport
connection. This was leaking open connections when running in
daemon mode. Might be a bug in Paramiko or their documentation
as they mention SFTPClient.close() should close the underlying
transport.

To fix, made transport an object variable so we can close it
ourself.

Signed-off-by: Rene Fragoso <ctrlrsf@gmail.com>